refactor cat tool to extract document reading part
